Implement room-geometry, render-cell, render-rect, render-line.
The new room-geometry function makes it simpler to get x, y, w, h data
for a given cell. The new render-rect and render-line functions make use
of room-geometry and anchor-point to output SVG. The new render-cell
takes a set of existing lines and declines to render ones which have
already been rendered.