Improve svg rendering and tests.
In particular, I was worried that certain errors were occurring because
the nested hashes in the render-env are not deeply merged, meaning that
a supplied hash could completely replace the previous one. This concern
may yet be founded, but it was not the cause of the errors. Keeping it
anyway.

Otherwise, added some tests to cover more SVG render options.